+++ R/editClusterST.R
@@ -5,21 +5,80 @@
 #' 
 #' Edit parameters and time series of an existing `st-storage` cluster (Antares studies >= v8.6.0).
 #' 
-#' @param area The area where to create the cluster.
-#' @param cluster_name Name for the cluster, it will prefixed by area name, unless you set `add_prefix = FALSE`.
-#' @param group Group of the cluster, one of : "PSP_open", "PSP_closed", "Pondage", "Battery", "Other". It corresponds to the type of stockage.
-#' @param storage_parameters Parameters to write in the Ini file. 
-#' @param PMAX_injection modulation of charging capacity on an 8760-hour basis. The values are float between 0 and 1.
-#' @param PMAX_withdrawal modulation of discharging capacity on an 8760-hour basis. The values are float between 0 and 1.
-#' @param inflows imposed withdrawals from the stock for other uses, The values are integer.
-#' @param lower_rule_curve This is the lower limit for filling the stock imposed each hour. The values are float between 0 and 1.
-#' @param upper_rule_curve This is the upper limit for filling the stock imposed each hour. The values are float between 0 and 1.
-#' @param add_prefix If `TRUE` (the default), `cluster_name` will be prefixed by area name.
-#' 
+#' @inheritParams createClusterST
 #' @template opts
 #' 
-#' @seealso [createClusterST()] to edit existing clusters, [removeClusterST()] to remove clusters.
+#' @note
+#' Put only properties or TS value you want to edit (see `examples` section).
 #' 
+#' @seealso [createClusterST()], [removeClusterST()]
+#' 
+#' @examples
+#' \dontrun{
+#' # study version >= "8.6.0"
+#'
+#' # edit an existing cluster (see doc approved groups)
+#' name_group <- "Pondage"
+#'
+#' editClusterST(area = "areaname", 
+#'               cluster_name = "clustername", 
+#'               group = name_group)
+#' 
+#' # edit properties
+#' all_params <- storage_values_default()
+#' all_params[["efficiency"]] <- 0.9
+#' all_params[["reservoircapacity"]] <- 1000
+#' all_params[["initiallevel"]] <- 0.5
+#' all_params[["withdrawalnominalcapacity"]] <- 250
+#' all_params[["injectionnominalcapacity"]] <- 200
+#' all_params[["initialleveloptim"]] <- TRUE
+#'
+#' editClusterST(area = "areaname", 
+#'               cluster_name = "clustername", 
+#'               storage_parameters = all_params)
+#'
+#' # edit time series
+#' inflow_data <- matrix(3, 8760)
+#' ratio_data <- matrix(0.7, 8760)
+#'
+#' editClusterST(area = "areaname", 
+#'               cluster_name = "clustername",
+#'               PMAX_withdrawal = ratio_data, 
+#'               inflows = inflow_data, 
+#'               PMAX_injection = ratio_data, 
+#'               lower_rule_curve = ratio_data, 
+#'               upper_rule_curve = ratio_data)
+#'
+#' # study version >= "9.2" (new parameters and TS)
+#' 
+#' # edit group (dynamic)
+#' name_group <- "MyOwnGroup"
+#'
+#' editClusterST(area = "areaname", 
+#'               cluster_name = "clustername", 
+#'               group = name_group)
+#'
+#' # edit properties
+#' my_parameters <- storage_values_default()
+#' my_parameters$efficiencywithdrawal <- 0.5
+#' my_parameters$`penalize-variation-injection` <- TRUE
+#' my_parameters$`penalize-variation-withdrawal` <- TRUE
+#'
+#' editClusterST(area = "areaname", 
+#'               cluster_name = "clustername", 
+#'               storage_parameters = my_parameters)
+#'
+#' # edit time series
+#' ratio_data <- matrix(0.7, 8760)
+#'
+#' editClusterST(area = "areaname", 
+#'               cluster_name = "clustername",
+#'               cost_injection = ratio_data, 
+#'               cost_withdrawal = ratio_data, 
+#'               cost_level = ratio_data, 
+#'               cost_variation_injection = ratio_data, 
+#'               cost_variation_withdrawal = ratio_data)
+#' }
 #' @export
 editClusterST <- function(area,
                           cluster_name, 
@@ -30,19 +89,31 @@ editClusterST <- function(area,
                           inflows = NULL,
                           lower_rule_curve = NULL,
                           upper_rule_curve = NULL,
+                          cost_injection = NULL,
+                          cost_withdrawal = NULL,
+                          cost_level = NULL,
+                          cost_variation_injection = NULL,
+                          cost_variation_withdrawal = NULL,
                           add_prefix = TRUE, 
                           opts = antaresRead::simOptions()) {
 
-  # basic checks
+  ## check study opts parameters ----
   assertthat::assert_that(inherits(opts, "simOptions"))
+  
+  ## check study version ----
   check_active_ST(opts, check_dir = TRUE)
+  
+  ##  check area ----
   check_area_name(area, opts)
   
-  # check 'group'
+  ## tolower area----
+  area <- tolower(area)
+  
+  ## check 'group'----
   .check_group_st(group = group, opts = opts)
   
   ##
-  # check parameters (ini file)
+  ## check parameters ----
   ##
   params_cluster <- NULL
   
@@ -59,21 +130,40 @@ editClusterST <- function(area,
     # check values parameters
     .st_mandatory_params(list_values = storage_parameters, opts = opts)
     
-    # check list of parameters
+    ### Standardize storage_parameters  ----
     params_cluster <- hyphenize_names(storage_parameters)
   }
   
-  # make list of parameters
-  area <- tolower(area)
-  if(!(is.null(params_cluster)&&is.null(group))){
-    cluster_name <- generate_cluster_name(area, cluster_name, add_prefix)
-    params_cluster <- c(list(name = cluster_name, group = group), 
-                        params_cluster)
-  }
+  ## Standardize cluster name + prefix ----
+  cluster_name <- generate_cluster_name(area, 
+                                          cluster_name, 
+                                          add_prefix)
+  
+  # all properties of cluster standardized
+  params_cluster <- c(list(name = cluster_name, 
+                           group = group),
+                      params_cluster)
+  
+  
+  
+  # delete it (for writeIni())
   if(is.null(group))
     params_cluster$group <- NULL
   
-  ##### API block ----
+  ## check dim data ----
+  
+  # According to Antares Version 
+    # default values associated with TS + .txt names files
+  list_local_values_params <- .default_values_st_TS(opts = opts)
+  
+  # check every ts parameter
+  for (name in names(list_local_values_params)){
+    if (!is.null(dim(get(name))))
+      if (!identical(dim(get(name)), c(8760L, 1L)))
+        stop(paste0("Input data for ", name, " must be 8760*1"))
+  }
+  
+  ## API block ----
   if (is_api_study(opts)) {
     # format name for API 
     cluster_name <- transform_name_to_id(cluster_name)
@@ -82,15 +172,16 @@ editClusterST <- function(area,
     # PATCH for properties 
     ##
     # adapt parameter names
-    list_properties <- list("group" = params_cluster[["group"]],
-                            "name" = cluster_name,
-                            "injectionNominalCapacity" = params_cluster[["injectionnominalcapacity"]],
-                            "withdrawalNominalCapacity" = params_cluster[["withdrawalnominalcapacity"]],
-                            "reservoirCapacity" = params_cluster[["reservoircapacity"]],
-                            "efficiency" = params_cluster[["efficiency"]],
-                            "initialLevel" = params_cluster[["initiallevel"]],
-                            "initialLevelOptim" = params_cluster[["initialleveloptim"]],
-                            "enabled" = params_cluster[["enabled"]])
+    list_properties <- list(
+      "group" = params_cluster[["group"]],
+      "name" = cluster_name,
+      "injectionNominalCapacity" = params_cluster[["injectionnominalcapacity"]],
+      "withdrawalNominalCapacity" = params_cluster[["withdrawalnominalcapacity"]],
+      "reservoirCapacity" = params_cluster[["reservoircapacity"]],
+      "efficiency" = params_cluster[["efficiency"]],
+      "initialLevel" = params_cluster[["initiallevel"]],
+      "initialLevelOptim" = params_cluster[["initialleveloptim"]],
+      "enabled" = params_cluster[["enabled"]])
     
     list_properties <- dropNulls(list_properties)
     
@@ -152,8 +243,8 @@ editClusterST <- function(area,
     }
     return(invisible(opts))
   }
-  #####-
   
+  ## write properties ----
   # path to ini file
   path_clusters_ini <- file.path(opts$inputPath, 
                                  "st-storage", 
@@ -164,13 +255,14 @@ editClusterST <- function(area,
     stop("'", cluster_name, "' in area '", area, "' doesn't seems to exist.")
   
   # only edition if parameters are no NULL
-  if(is.null(params_cluster))
+  is_null_parameter <- all(names(params_cluster)%in%"name")
+  if(is_null_parameter)
     warning("No edition for 'list.ini' file", call. = FALSE)
   else{
     # read previous content of ini
     previous_params <- readIniFile(file = path_clusters_ini)
     
-    if (!tolower(cluster_name) %in% tolower(names(previous_params)))
+    if (!cluster_name %in% tolower(names(previous_params)))
       stop(
         "'", 
         cluster_name, 
@@ -181,7 +273,7 @@ editClusterST <- function(area,
     
     # select existing cluster
     ind_cluster <- which(tolower(names(previous_params)) %in% 
-                           tolower(cluster_name))[1]
+                           cluster_name)[1]
     previous_params[[ind_cluster]] <- utils::modifyList(x = previous_params[[ind_cluster]], 
                                                         val = params_cluster)
     names(previous_params)[[ind_cluster]] <- cluster_name
@@ -194,61 +286,50 @@ editClusterST <- function(area,
     )
   }
   
-  ##
-  # check DATA (series/)
-  ##
   
-  # datas associated with cluster
+  ## write TS ----
+  
+  ##
+  # Write TS PART ("series/")
+  ##
+
+  # Path folder for TS
   path_txt_file <- file.path(opts$inputPath, 
                              "st-storage", 
                              "series", 
-                             tolower(area), 
-                             tolower(cluster_name))
+                             area, 
+                             cluster_name)
   
-  # PMAX_injection
-  if(!is.null(PMAX_injection)){
-    fwrite(
-      x = PMAX_injection, row.names = FALSE, col.names = FALSE, sep = "\t",
-      file = file.path(path_txt_file, 
-                       paste0("PMAX-injection", ".txt"))
-    )
-  }
+  # list of params of TS 
+  list_local_params <- list(PMAX_injection = PMAX_injection,
+                         PMAX_withdrawal =  PMAX_withdrawal,
+                         inflows =  inflows ,
+                         lower_rule_curve =  lower_rule_curve,
+                         upper_rule_curve =  upper_rule_curve,
+                         cost_injection =  cost_injection,
+                         cost_withdrawal =  cost_withdrawal,
+                         cost_level =  cost_level,
+                         cost_variation_injection =  cost_variation_injection,
+                         cost_variation_withdrawal =  cost_variation_withdrawal)
   
-  # PMAX_withdrawal
-  if(!is.null(PMAX_withdrawal)){
-    fwrite(
-      x = PMAX_withdrawal, row.names = FALSE, col.names = FALSE, sep = "\t",
-      file = file.path(path_txt_file, 
-                       paste0("PMAX-withdrawal", ".txt"))
-    )
-  }
-  
-  # inflows
-  if(!is.null(inflows)){
-    fwrite(
-      x = inflows, row.names = FALSE, col.names = FALSE, sep = "\t",
-      file = file.path(path_txt_file, 
-                       paste0("inflows", ".txt"))
-    )
-  }
-  
-  # lower_rule_curve
-  if(!is.null(lower_rule_curve)){
-    fwrite(
-      x = lower_rule_curve, row.names = FALSE, col.names = FALSE, sep = "\t",
-      file = file.path(path_txt_file, 
-                       paste0("lower-rule-curve", ".txt"))
-    )
-  }
-  
-  # upper_rule_curve 
-  if(!is.null(upper_rule_curve)){
-    fwrite(
-      x = upper_rule_curve, row.names = FALSE, col.names = FALSE, sep = "\t",
-      file = file.path(path_txt_file, 
-                       paste0("upper-rule-curve", ".txt"))
-    )
-  }
+  # write TS
+  lapply(names(list_local_params), function(x_val){
+    if(!is.null(list_local_params[[x_val]])){
+      # name file
+      name_file <- list_local_values_params[[x_val]][["string"]]
+      # write disk
+      dt_to_write <- as.data.table(list_local_params[[x_val]])
+      fwrite(
+        x = dt_to_write, 
+        row.names = FALSE, 
+        col.names = FALSE, 
+        sep = "\t",
+        file = file.path(path_txt_file, 
+                         paste0(name_file, 
+                                ".txt"))
+      )
+    }
+  })
   
   # Maj simulation
   suppressWarnings({
@@ -256,5 +337,4 @@ editClusterST <- function(area,
   })
   
   invisible(res)
-  
 }