+++ NEWS.md
@@ -3,6 +3,12 @@
 # antaresEditObject 0.9.2.9000
 (cf. Antares v9.2 changelog)
 
+NEW FEATURES :  
+
+* `createClusterST()`/`editClusterST()` :  
+  - **New properties** (*efficiencywithdrawal*, *penalize-variation-injection*, *penalize-variation-withdrawal*, see list of properties according to study version of Antares with `storage_values_default()`)  
+  - **New optional time series** (cost-injection, cost-withdrawal, cost-level, cost-variation-injection, cost-variation-withdrawal)
+
 ### Breaking changes  :  
   - `createClusterST()`/`editClusterST()` : parameter `group` is now dynamic and have no restriction  
   - `createClusterST()` : for a study < 9.2, execution will be STOP if `group` is not included in list (see doc)  
@@ -10,7 +16,7 @@
 
 ### DOC :  
 
-A new article exposing new features of Antares Simulator v9.2 is available [here](https://rte-antares-rpackage.github.io/antaresEditObject/release920/dev/articles/Antares_new_features_v92.html)
+A new article exposing new features of Antares Simulator v9.2 is available [here](https://rte-antares-rpackage.github.io/antaresEditObject/release920/dev/articles/Antares_new_features_v920.html)